LayoutTests/storage/indexeddb/modern/deletedatabase-request-event.html

This test calls deleteDatabase on window.indexedDB and verifies that an event is fired at the request.<br>
While at this stage in modern IndexedDB development, that event is hardcoded to be an error.<br>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

var request = window.indexedDB.deleteDatabase("TestDatabase");

request.onsuccess = function()
{
	alert("success");
	done();
}
request.onerror = function(e)
{
	alert("error " + e);
	done();
}

alert(request);

</script>
